Create Build_Main_DG-LDMOS_wSTI.in
This script creates the structure that is the focus of the project using Silvaco Athena. It is considered "the default" for this project. It was produced by trying to emulate conventional LDMOS and DG-LDMOS, then manually changing parameters (and the addition of a shallow isolation trench) until a somewhat decent result was achieved. The design is not optimal, but it is what I was able to produce within the time allotted and other constraints of the project.